requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
portalocker>=2.8.0
httpx>=0.27.0
nicegui>=1.4.0